Refactor storage into feature modules
